+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/service/delegates/communications/ServerCommunications.java
@@ -1,20 +1,21 @@
 /*
+ * Copyright (c) 2020 Broadcom.
  *
- *  Copyright (c) 2019 Broadcom.
- *  The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
+ * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
- *  This program and the accompanying materials are made
- *  available under the terms of the Eclipse Public License 2.0
- *  which is available at https://www.eclipse.org/legal/epl-2.0/
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
  *
- *  SPDX-License-Identifier: EPL-2.0
+ * SPDX-License-Identifier: EPL-2.0
  *
- *  Contributors:
- *    Broadcom, Inc. - initial API and implementation
+ * Contributors:
+ * Broadcom, Inc. - initial API and implementation
+ *
  *
  */
 
-package com.ca.lsp.cobol.service.delegates;
+package com.ca.lsp.cobol.service.delegates.communications;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
