+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/service/delegates/formations/Formations.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019 Broadcom.
+ * Copyright (c) 2020 Broadcom.
+ *
  * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
  * This program and the accompanying materials are made
@@ -9,9 +10,11 @@
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
- *   Broadcom, Inc. - initial API and implementation
+ * Broadcom, Inc. - initial API and implementation
+ *
+ *
  */
-package com.ca.lsp.cobol.service.delegates;
+package com.ca.lsp.cobol.service.delegates.formations;
 
 import com.ca.lsp.cobol.service.MyDocumentModel;
 import lombok.experimental.UtilityClass;
