+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/TestModule.java
@@ -19,8 +19,8 @@ import com.broadcom.lsp.cdi.module.DefaultModule;
 import com.ca.lsp.cobol.positive.CobolTextRegistry;
 import com.ca.lsp.cobol.positive.ZipTextRegistry;
 import com.ca.lsp.cobol.service.MyTextDocumentService;
-import com.ca.lsp.cobol.service.delegates.Communications;
-import com.ca.lsp.cobol.service.delegates.ServerCommunications;
+import com.ca.lsp.cobol.service.delegates.communications.Communications;
+import com.ca.lsp.cobol.service.delegates.communications.ServerCommunications;
 import com.ca.lsp.cobol.service.delegates.validations.CobolLanguageEngineFacade;
 import com.ca.lsp.cobol.service.delegates.validations.LanguageEngineFacade;
 import com.ca.lsp.cobol.service.mocks.MockWorkspaceService;
